Module p3_backend::models::employee

Structs

Employee

Representation of Employees in the database.

FromRequestError

Custom error type needed to implement FromRequest for Employee.

get_employees

Returns a JSON array of employee objects.

post_employees

Inserts the posted employee into the database. If id is None, it will be generated by the database.

put_employees

Updates the Employee in the database.

user_from_token

Uses the Employee extractor to get it from the values in the JWT in the header. Returns the employee as a JSON object.